@@ -10,32 +10,37 @@
     "types": "tsc -p ."
   },
   "dependencies": {
-    "@intlify/eslint-plugin-vue-i18n": "^2.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.8.0",
-    "@typescript-eslint/parser": "^7.8.0",
-    "@typescript-eslint/utils": "^7.8.0",
-    "eslint": "^8.57.0",
+    "@eslint-community/eslint-plugin-eslint-comments": "^4.4.0",
+    "@eslint/js": "^9.12.0",
+    "@intlify/eslint-plugin-vue-i18n": "^3.0.0",
+    "@types/eslint-config-prettier": "^6.11.3",
+    "@types/eslint__js": "^8.42.3",
+    "@typescript-eslint/eslint-plugin": "^8.8.1",
+    "@typescript-eslint/parser": "^8.8.1",
+    "@typescript-eslint/utils": "^8.8.1",
+    "@vitest/eslint-plugin": "^1.1.7",
+    "eslint": "^9.12.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-import-resolver-typescript": "^3.6.3",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jsonc": "^2.15.1",
-    "eslint-plugin-playwright": "^1.6.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsonc": "^2.16.0",
+    "eslint-plugin-playwright": "^1.7.0",
     "eslint-plugin-tsdoc": "^0.3.0",
-    "eslint-plugin-unicorn": "^52.0.0",
-    "eslint-plugin-vitest": "^0.3.18",
-    "eslint-plugin-vue": "^9.25.0",
-    "eslint-plugin-vuejs-accessibility": "^2.3.0",
+    "eslint-plugin-unicorn": "^56.0.0",
+    "eslint-plugin-vue": "^9.29.0",
+    "eslint-plugin-vuejs-accessibility": "^2.4.1",
     "jsonc-eslint-parser": "^2.4.0",
-    "typescript": "^5.2.2",
-    "vue-eslint-parser": "^9.4.2"
+    "typescript": "5.6.3",
+    "typescript-eslint": "^8.8.1",
+    "vue-eslint-parser": "^9.4.3"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3.0.2",
+    "@eslint/eslintrc": "^3.1.0",
     "@swc/cli": "^0.4.0",
-    "@swc/core": "^1.3.82",
-    "@typescript-eslint/rule-tester": "^7.8.0",
+    "@swc/core": "^1.7.35",
+    "@typescript-eslint/rule-tester": "^8.8.1",
     "babel-plugin-add-module-exports": "^1.0.4",
-    "vitest": "^2.0.5"
+    "vitest": "^2.1.2"
   }
 }
